Fixed book row dblclick; always refresh rdrs/bks.
Double-clicking was crashing because it was trying to load a
non-existant agent with the book id instead of a good.

Visualization redraw skipping when the visualization tab wasn't active
was breaking things with the new controls-above-the-tabs layout, since
you can now run/step from the readers or books tabs--but because the
visualization tab wasn't active, reader and book lists weren't being
updated.