Added full constructor; made default less variable
The previous non-informative model parameters were both highly
non-informative *and* highly variable; change the latter so that the
parameters are N(0,1) for the first K-1 parameters and N(1,1) for the
last one (instead of N(0, LARGE) for all of them.  n=1e-6 is preserved,
however, so that when using it as a prior it is still weak, but not as
variable.

Added a full model parameters constructor (which gets invoked by the
above constructor).