Reordered drawing order to make display cleaner
This helps particularly when there are larger number of readers/books:
the display was basically useless before because the circles and lines
cover everything.

Now lines are drawn first, then books, then readers, then the axes.

Also removed the circle for number of new books (the number of lines
already shows that).