Moved belief params to Creativity.parameters
They were previously hardcoded in Creativity.setup() (and before that,
hard-coded in gui.cpp).  This moves them to the publically accessible
Creativity.parameters struct.

This also adds static parameters() methods to
belief::{Demand,Profit,Quality} so that the number of parameters
doesn't need to be hard-coded.  (Eigen will complain about the beta
initializations having the wrong number of fields if the model sizes
change, but that's a good thing).