Increased the prior_n for noninformative models
The 1e-6 was failing catastrophically if attempting to draw() from the
noninformative model because getting the t-distributed value requires
getting a chi-squared(n) value, but for 1e-6, such draws are usually
values below machine epsilon.  Even 1e-3 still produces <epsilon draws
about 2.4% of the time.  1e-1 and 1e-2 seemed okay, but just to be on
the safe side, I bumped this up to 0.5.

Hopefully that doesn't make it too "informative", but coupled with the
large noninformative s^2 value of 1e+6, I think this will still be fine.