Reformulate demand belief (now has P,q,P^2,q^2)
The demand belief previously had P^D and q^D terms (though not quite:
the exponentiation on q was sign-preserving because q could be
negative, which is no longer possible).

The two terms have now become 4: P, P^2, q, q^2 (in the 2d case, this
means adding linear terms, in the 1-d case, this is adding squares, and
in other cases, the models are non-nested).