Added visualization options to GUI Settings tab.
The visualization gets updated and redrawn immediately when the options
are changed (and cached visualization stages cleared).

The stored R-trees are also affected: they now only contain visible GUI
readers/books, and are recalculated when the setting is changed.  Since
it's also now quite possible to have empty rtrees (i.e. when all of
readers, live books and dead books are disabled), the rtree is now
wrapped in a unique_ptr so as to be able to distinguish between "rtree
not set yet" and "empty rtree"--recalculation is need in the former
case but not the latter.