Rearranged book list columns
This order is more logical: "Life" moves from the end to right after
"Created", and "Quality" moves one to the right (so that it's now
between Mkt? and Price).

I almost took Copies out, since it's just Sales+Pirated, but then
though it might still be useful to have for sorting purposes.