Added V matrices; report s2; replaced macros.
The macros happening here were insane; replaced them with sane methods
instead.

Added ability to add a set of matrix labels to a grid.  This uses a
(r+1)*(c+1) set of grid labels; the first row and column is used for
labels, the remaining is the matrix element labels.

This is now used to display the V matrix for the beliefs.