Added creationEffort(); creation_scale 1 -> 10
creationEffort() is the inverse of creationQuality()--that is, it gives
an effort level required to hit a target quality level.

Also added a scratch test script to test it.

creation_scale was 1.0, but that's much too small: changed it to 10 (so
that feasible quality levels are 0-69 instead of 0-6.9 (the latter is
rather unlikely: it would require spending 1000 income).