Reader step/creation settings; changed creation fn
This adds settings for reader step sd, creation shape, and U[] limits
for creation_scale.

This also changes the creation function from:

    q(l) = alpha [(l+1)^(1-beta) - 1] / (1-beta)

to:

    q(l) = alpha [(l+1)^beta - 1] / beta

i.e. newbeta = 1-oldbeta.  This makes the function a bit simpler.  It
also means beta=0 (instead of beta=1) corresponds to ln(), and the new
beta restriction (for concavity) is beta < 1 instead of beta > 0.