FileStorage: Remove belief dedupe compression
With weakening, deduplication makes much less difference than it used to
(because a reader's beliefs always change from one period to the next
there is no longer much temporal compression).  Dedupe compression still
made a small difference, but presumably that was all in the initial
periods where reader observations are identical.

This ought to make file storage considerably faster as well as now the
belief can just be written, while before it had to be checksummed, then,
if a matching checksum was found, it had to be read from disk and
compared (to prevent checksum collisions).  Reading may be a bit faster
(depending on OS caching and system hardware) as reading a belief no
longer involves seeking to some other point in the file then seeking
back.

This change makes noninformative and default-constructed beliefs smaller
(1 byte each instead of 8), and makes (non-duplicated) beliefs smaller
(1 byte instead of 12 [8 for the "immediately follows" location value,
and 4 for the K value]).  This limits K values to 120, but since the K
values in use are all less than 10, that seems like not a big deal.