upperBounds()[i] -> upperBound(i) etc.
The subscripting and extra proxy object was pointless verbose.

This also adds const versions of upperBound() and lowerBound() so that
querying (i.e. restricted() and operator double()) can be done on a
const LinearRestricted.

The actual code for restricted() and operator double() is now moved
into Linear (hasRestriction() and getRestriction()), with the proxy
object simply calling those methods.