update() and weaken() via B instead of Linear
The base class may have some override, so it seems more correct to call
B::update() and B::weaken() instead of Linear::update() and
Linear::weaken().

Currently only Linear and LinearRestricted are being used as bases, and
so this makes no difference (because LinearRestricted doesn't override
them), but this is more correct.