Always use a tempfile; default to same dir.
Using a tempfile always is a good idea (that way the final file is
moved into place only if the simulation finishes, but not if it is
aborted for some reason).  If --tmpdir isn't specified, it does in the
same directory as the output file (so that the rename will likely
succeed, avoiding the need to copy).