Default to threaded output
Even with the same seed, results aren't reproducible, so defaulting to
off doesn't make a lot of sense in that respect.

Furthermore, threading now scales very well (at least up to available
CPUs/threads), so default the GUI to use threads.  (Since the cli is
more likely to be run in a loop with several running at the same time,
don't default to threads there--but when the GUI is used to run a
simulation, it's much more likely to be the only thing running).