Fix segfault when going back with book selected
If you selected a book, then moved back to a time when the selected
book's location in the list doesn't exist in the previous time, you'd
get a segfault because the "hint" location wasn't checked against being
larger than the maximum hint location.  Fixed so that a passed-in hint
position > max gets truncated to max.