Noninformative value fixes
V itself, not s2*V, is what needs to be big for a noninformative model,
because V^-1 gets added to XtX--but if V is I, it will be "big", which
is bad (it was massively biasing the estimates!).

s2 being small (now 1.0) hopefully isn't a problem, because it only
enters through s2_*n_, and n_ is small.  (n_ could be smaller still, I
think, without being a problem).