Added std::string operator; removed print method
It's more versatile to be able to convert to a string than just change
how a Linear object gets print, so made that change (operator<< now
print the string conversion instead of calling print()).

This also renames the print_name() virtual to display_name(), and adds
display_name overrides to Demand/Quality/etc.