drawGibbs t drawing now works well
Drawing now works by drawing a independent N(0, I) vector, z, and then
alternates between drawing sigma values (truncated suitably so that
beta_ + sigma*s*Ainv*z satisfies the constraints) and new z_i values
(truncated to satisfy the constraints under sigma and the other z_{-i}
values).  Although this seems to be equivalent to what was being done
before, it apparently isn't because this doesn't exhibit the bias seen
in the previous implementation.  (It may be that, as per Hofert (2013)
"On Sampling from the Multivariate t Distribution", we were actually
erroneously multiplying the mean by A as well--but I didn't investigate
in great detail whether that was actually happening).

This commit also removes truncNorm(), truncGamma() in favour of a
generic truncDist replacement that takes a boost distribution object
(from which it can lookup cdf, cdf complement, quantile, and quantile
complements), instead of programming those directly here.