Fixes for development and future Eigen versions
The define guards for pre-Eigen-3.3 code weren't actually right: needed
constructors and assignment operatores were being omitted for Eigen 3.3+
code.

This also allows the current Eigen development version (3.2.90) to be
used with the efficient move constructors (as long as it has the RVALUE
definition, which got added to Eigen-dev at the same time as the move
constructor fixes).