Make sure XtX is invertible; s2 value fix
XtX needs to be invertible; this is *probably* true when X has full
column rank, but since we'll need to calculate XtX anyway, might as
well do the full check.

s2 was wrong, because it used Vinv() and beta_diff, but both of those
are invalid when loading full-rank data for the first time.