Nixed fullyInformative(); added beta_names_
noninformative is repurposed slightly: it stays true as long as the
model hasn't been able to incorporate the data yet (i.e. because XtX is
not invertible)

This also switches the inverses to fullPivHouseholderQr instead of
colPivHouseholderQr, which hopefully delivers more accurate results.

model.names() can now be used to get/set the names of the model; these
are used in InfoWindow, and also show up if the model is printed.