Don't store noninf data.
With the unweakened noninf data, it's getting to be too much: just
store such beliefs as noninformative instead.  It isn't as if anything
READING a state needs the noninformative data, and though I've thought
of the ability of "resuming" a simulation, that currently doesn't
exist, and I'm not sure it's worth bothering with.