InfoWindow inheritance: (Reader|Book)InfoWindow
Handling both reader and book details inside InfoWindow was sort of
dumb: the refresh method was basically a a big if/else, and the two
constructors were just one for readers and one for books.

This is now split up (with inheritance): InfoWindow is the abstract base
class, providing all the general purpose functions (displaying fields,
matrices, etc.), while ReaderInfoWindow and BookInfoWindow handle the
reader-specific and Book-specific details.