Added support for 1-D simulations in GUI
1D simulations are circular cities, and show up as such.

3+ dimensions are also sort of supported: only the first two dimensions
are displayed (but the full dimensionality is visible in detailed
fields).  3+ dimensional simulations can't actually be started for now,
as the GUI has no way to configure which dimensions are displayed
(though that would be easy enough to add by just changing the
design.subdimensionX and design.subdimensionY variables).

Bump eris requirement to (upcoming) 0.3.2 for required Position changes.