Reverse scroll direction on graph area for new Gtk
Gtk 3.16.1 changed the scroll direction for horizontal sliders, and
made it very strange: use the scroll wheel on the slider moved one way,
scrolling on the graph area moved the other way.  This change makes
scrolling on the graph area move the same way as on the slider (based
on the Gtk version in use).