Save as support for saving a running/loaded sim
Previously the GUI (and API) didn't allow writing a simulation to a
file if the simulation had already been initialized or loaded from a
file.  This commit allows that, and moreover copies any current states
into the saved file.

From a UI point of view, this means two things are now possible:
running a simulation in-memory, then deciding the save it; and loading
a simulation then saving it somewhere else (though the latter is
probably not very useful since you could just copy the file).

This also changes the structure of states a little bit to use (ordered)
map and set objects instead of unordered_map and unordered_set, so that
reading then rewriting a simulation is expected to produce identical
output.