Fix data --precision maximum/default
digits10 was being used by default (and the maximum) for --precision,
but that's only the number of digits that survives a round-trip through
conversion from decimal to double; instead this should be max_digits10,
which is the number of decimal digits required to distinguish between
all possible double values after conversion to decimal.  (In other
words, it should be (and was documented as such; and now is) the number
of digits to represent all significant information in a double).