SimpleVar: don't accept Ref<VectorXd>
Accepting a Ref<VectorXd> can end up accepting a temporary in various
cases, and by the time we use it, the temporary is gone.  This changes
SimpleVariable to store a Block instead, and to take an actual lvalue
VectorXd or a MatrixXd and column number, which won't compile in cases
where a temporary would be created.

This limits it considerably (i.e. you can't pass arbitrary things: only
a plain VectorXd or a column of a MatrixXd) but that seems an
acceptable limitation for the use case.