Add libcreativity-data library
Everything in creativity/data/ plus the relevant cmdargs/ cpp files now
get compiled into a new libcreativity-data.so library (instead of
libcreativity.so), and all of it can be disabled with the new DATA
option.

This also renames libcreativitygui to libcreativity-gui, and moves the
cmdargs/GUI.cpp into it instead of libcreativity.so.