Added preliminary results parser
creativity-results takes the CSV results as produced by
creativity-data, and will (once finished) produce tabulated output of
the relationships in the variables.

Currently this parses the CSV files (which were also simplified a bit,
to eliminate things like quoting, which we really don't need here).
Next is to use SUR to actually analyze that data.