Remove more default values
Having the default values in gui.glade is a bad idea: it's pointless and
deceptive, because the defaults get set from the CreativitySettings
object regardless of the gui.glade value.  Moreover, not having the
defaults is better in the sense that it will quickly reveal values that
don't have the default being set properly.