Initial behaviour: determine effort, not quality
The initial behaviour drawing quality was a big problem: a quality draw
of, say, 10 combined with a scale of 1 would require effort of 22025,
which is obviously impossible.  In the previous code, a check for
effort <= income prevented such a book from being created, but the
addition of creation_fixed broke this (not taking creation_fixed into
account) and exposed the nonsensical effort values.

This commit changes the initial behaviour specification by drawing
*effort* from a uniform distribution (U[0,100], by default) and
creating with the resulting quality.

This also fixes a rather serious seeming bug with market pricing being
improperly handled, exhausting all effort by recalculating the price
for the most profitable book repeatedly.  I *think* this means no
author would ever leave more than 1 book on the market (except in the
unusual case that 2 books have identical predicted profitability).