Reader: utility function penalty overhaul
Previously, the "u_poly_" was a decreasing function of distance that
started out positive.  This didn't make a huge amount of sense:
distance is a bad thing, so it makes more sense to model it as a
penalty function (just like the number-of-books penalty).  Thus the
net utility of a book is the quality of a book minus the distance
penalty, minus the price/cost, minus the number-of-books penalty.

Renamed the "utility polynomial" to "distance penalty polynomial" (while
also making it negative) and renamed "penalty" to "number of books
penalty".

This changes the default distance penalty to d + 0.25d^2, while
previously it was (effectively) -4 + d.

This also increases the number of books penalty to n^2 (from 0.25 n^2).