Added smarter minimum effort value for creation
It isn't rational to produce something at quality 0, because there is
always a non-zero cost to obtain a book: both the price, *and* the book
size penalty and almost always some distance penalty.

This changes the profit maximization to accept a minimum effort
(instead of using 0), and has readers use the minimum of the effort
level required to create a book of quality that just clears the
would-never-buy threshold of potential readers.