Revenue and profit: include prize money, cost
Revenue was only calculating private market revenue instead of total
revenue.

Profit was doing the same, and was also not including the fixed creation
cost (just the effort component).

This also removes the duplicated comments from the .cpp: duplication is
bad, but also some may be out of date and thus conflicting with the
.hpp values, which is worse.