Remove "optimizations"
Compiling with -march=native -O3 (which turns on SSE versions > 2 and AVX)
actually ends up taking *double* the time to run the first hundred iterations
of creativity-cli compared to compiling with just -O2 (just turning off
-march=native but leaving -O3 instead of -O2 makes things about 10% slower).

Moreover something turned on there seems to be resulting in in
non-repeatability of random numbers that has been happening: this commit seems
to fix that problem: the same seed now generates the same result, which is
fantastic!

(This whole thing gives me unpleasant flashbacks of years ago when I used
Gentoo instead of Debian).