Fixed book author line problems
Author lines weren't showing up properly (according to configured
settings), and were being drawn incorrectly: this was partly due to
lines not being drawn at all unless on and off-market were enabled (and
hence public lines required at least one of the others to be enabled).

Then there was an "if" and should have been an "else if" that caused
all private market books to be drawn in the line style of off-market
books.