Changed 'struct RGBA' -> 'class RGBA final'
This is purely cosmetic, of course.  RGBA started as a simple struct for
4 double values with a few methods, but it's become much larger since
then with various methods; it seems conceptually more of a class than a
struct now, so this commit renames it appropriately.

Also added final since it's unlikely to be inherited from (and final
might let the compiler optimize slightly better).