Axis labels; right axis ticks; more legend pos.
1. Axis labels.  The y-axis label can be rotated (default) or not, and
is on the same side of the graph as the ticks (see next item).

2. y-axis ticks can be drawn on the right (left is still default).

3. Legend positioning is now much more capable, and simpler.  The
position is now a general position: either none, inside, or one of the
four sides (for outside); two new variables control where it ends up on
a scale relative to the graph size (either inner, for inside, or outer,
for the outside ones).  When on a side with stuff on it (ticks, label,
whatever), the legend goes outside the stuff except for the top, where
it goes between the title and the graph.

4. New method to update all the fonts at once with whatever is set in a
new FontDescription.  Another method takes a string to change just the
font family (but not size, slant, weight, etc.)