CSVParser & quantile: Eigen vector -> stl vector
CSVParser is sort of a nuissance when I want stl containers, so change
it to use std::vector<double> instead of an Eigen::RowVectorXd.

quantile also gets changed to accept a more stl-style iterator pair
in addition to the current std::vector or eigen vector (the former now
calls the pair version instead of the eigen version via an Eigen::Map).