Fix storage bug: don't hold Creativity shared ptr
Reader, Book, etc. stored shared_ptrs to the Creativity object owning
their simulation, but this means:

Creativity -> Simulation -> Reader
     ^------------------------'

and so resetting the Creativity shared_ptr (in creativity->cli) wasn't
actually deleting the object.

This commit changes all the members (Reader, Book, etc.) to store a hard
reference (either Creativity& or const Creativity&)--since the members
belong to the simulation, which is destroyed during the Creativity
destruction.

This was causing a serious error in storage, because the lack of
destruction meant the file wasn't closed.  This was compounded by
flush_for not actually doing a file handle flush (since that was
handled in FileStorage's flush override), though that has been fixed in
the previous commit.

With that gone, there really is no need to force Creativity access
through a shared_ptr, and so this commit also makes Creativity used as
a regular class just about everywhere (except cli, which uses a
unique_ptr--because it needs to control destruction).