GUI: use MRU caching on the gui elements/graph
Loading a simulation in the GUI took a huge amount of memory (once
having viewed every single state, i.e. so the cache is fully loaded).
For example, with a 1600x1300 pixel window, a typical 300-period
simulation took around 4.5GB of memory once every single state has been
loaded and cache populated.

This commit adds most-recently-used caching of the ten
most-recently-viewed states, so that flipping back and forth within a
set of up to 10 states is fast, but outside that range the GUI state data
and graph have to be redrawn.

This reduces the (maximum) memory requirement on the same
simulation/window size tested above to about 188MB.  The downside is
that you can't scroll wildly through the simulation to see the "action",
but I'm not sure that was all that useful a feature anyway, given the
memory requirement (especially because, in order to work properly, you
had to scroll either slowly or many times to get the cache fully
populated).