Make run-sim use same dir as typical-sim; .xz
run-sim was using a full timestamp instead of just the date, which is a
little annoying; changed it to use the date~~tag format, like
typical-sim.  It also now puts its files into a 'randomized'
subdirectory, instead of dropping them directly into timestamp~~tag.

Updated the help output of both to reflect that crstate files are
compressed now.