Use mkt books over creation cycle, not previous period
Many simulations have very strong creation cycles, especially in
pre-piracy periods, due to the use of a single lag: if creation_time is
1, for instance, there could exist an equilibrium of
(high,low,high,low,high,low,...) because the high number of books in the
previous period induces low creation undertaking this period, then the
resulting low mkt books in the next period induces high output in the
following one.

This fix is an attempt to address the cyclical behaviour by smoothing
out the belief to using the average over the last `creation_time+1`
periods.

This is unlikely to eliminate all cycles, however: some cycles happen
because talented creators create as often as possible, and because
beliefs begin at the same time, they are all following the same action.