Add some randomness to creation_time
Attempt to address cyclical behaviour, part 2.

One reason for cyclical behaviour is that the top creators all want to
create as much as possible, but when creation_time > 0 this means they
are cycling and all end up creating at the same time (this is amplified
by the fact that they start using beliefs at the same time).

This adds some randomness to the creation time, so that actual creation
time is creation_time + [-1,0,1] with equal probability.  (This only has
an effect when creation_time >= 1, of course, because we can't apply a
mean-preserving spread while maintaining mean 0--and anyway the cyclical
behaviour won't happen when people can create every single period).