Avoid possible insufficient assets for tax
Authors don't take the market cost into account when considering the effort limits if
creation_time is > 0, but if creation_time == 1, it's possible that
release happens immediately; in that case, the (unexpected) market cost
being removed from assets could result in assets not being sufficient to
cover due taxes, which would then hit a negative bundle error.

This fixes it by delaying the release for a period if current assets are
not sufficient to both cover the market cost and pay taxes (since the
author won't be creating next period, they should have funds available
at that point to cover both).