Add `Policy' class to handle policy operations
Manually playing with bits in various places is dirty: this adds a new
Policy class that abstracts all the bit values, while providing various
nicer operations and methods related to a policy value.

This also includes a eris::serialize::serializer<Policy> implementation
(which is just a simple wrapper around a serializer<uint32_t>) so that
a Policy object can be directly serialized.